Moved randomgenerator to pkg
